pyproject.toml
[project]
name = "ragproductionapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.1",
    "google-generativeai>=0.8.6",
    "groq>=1.2.0",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.21",
    "llama-index-readers-file>=0.6.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.57.0",
    "uvicorn>=0.46.0",
]
